minor: Revamp layout again
